admin: separate ui dashboard and data hooks
<!-- ps-id: 436bda9a-3df3-4962-bf06-698f669b965e -->
