+++ packages/admin-config/src/dashboards/definitions/color.dashboard.ts
