+++ packages/admin-config/src/dashboards/definitions/post.dashboard.ts
