+++ packages/admin-config/src/dashboards/definitions/product.dashboard.ts
