+++ packages/admin-config/src/hooks/hook.types.ts
@@ -1,0 +1,5 @@
+export type HookConfig<T> = {
+  name: string;
+  modelName: string;
+  beforeCreate: (record: T) => object | Promise<object>;
+};
