+++ packages/admin-config/src/hooks/index.ts
@@ -1,0 +1,28 @@
+/**
+ * Hooks allow you to define custom logic that should run at specific points in the lifecycle of a model's record
+ * (e.g., beforeCreate, beforeDelete). This separation helps organize custom business logic while keeping it
+ * clean and reusable.
+ *
+ * Each model can have its own hooks file where specific logic is implemented, ensuring a clear separation of concerns.
+ */
+import { HookConfig } from './hook.types';
+import { UserHooks } from './definitions/user.hooks';
+
+export function getHookConfigs(): HookConfig<unknown>[] {
+  return [UserHooks];
+}
+
+export function getHookConfig(modelName: string): HookConfig<unknown> {
+  const hookConfigs = getHookConfigs();
+  const hookConfig = hookConfigs.find(
+    (d) => d.modelName.toLowerCase() === modelName.toLowerCase(),
+  );
+
+  const NullSafeHookConfig: HookConfig<unknown> = {
+    name: modelName,
+    modelName,
+    beforeCreate: async (record: unknown) => record,
+  };
+
+  return hookConfig ?? NullSafeHookConfig;
+}
